# openapi.yaml
openapi: 3.0.0
info:
  title: WhatsApp Automation Bot API
  description: Production-ready WhatsApp automation bot system with API-like features powered by Playwright
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /whatsapp/send:
    post:
      tags:
        - WhatsApp
      summary: Send message
      description: Send a text or media message to a recipient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recipient
              properties:
                recipient:
                  type: string
                  description: Phone number or contact name
                  example: "1234567890"
                content:
                  type: string
                  description: Message content (required if mediaUrl not provided)
                  example: "Hello from WhatsApp Bot!"
                mediaUrl:
                  type: string
                  format: uri
                  description: URL of media file (required if content not provided)
                  example: "https://example.com/image.jpg"
                mediaType:
                  type: string
                  description: Type of media (required if mediaUrl provided)
                  enum: [image, video, audio, document, sticker]
                  example: "image"
                priority:
                  type: string
                  description: Message priority
                  enum: [low, normal, high]
                  default: "normal"
                  example: "normal"
      responses:
        '200':
          description: Message queued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Message queued successfully"
                  data:
                    type: object
                    properties:
                      jobId:
                        type: string
                        example: "550e8400-e29b-41d4-a716-446655440000"
                      status:
                        type: string
                        example: "queued"
                      estimatedProcessingTime:
                        type: string
                        format: date-time
                        example: "2024-01-01T12:00:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /whatsapp/send/text:
    post:
      tags:
        - WhatsApp
      summary: Send text message
      description: Send a text message to a recipient (simplified endpoint)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recipient
                - content
              properties:
                recipient:
                  type: string
                  description: Phone number or contact name
                  example: "1234567890"
                content:
                  type: string
                  description: Message content
                  example: "Hello from WhatsApp Bot!"
                priority:
                  type: string
                  description: Message priority
                  enum: [low, normal, high]
                  default: "normal"
      responses:
        '200':
          description: Text message queued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Text message queued successfully"
                  data:
                    type: object
                    properties:
                      jobId:
                        type: string
                      status:
                        type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /whatsapp/qr:
    get:
      tags:
        - WhatsApp
      summary: Get QR code
      description: Get QR code for WhatsApp login
      responses:
        '200':
          description: QR code retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: string
                    format: byte
                    description: Base64 encoded QR code image
                    example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
        '404':
          description: QR code not available
        '401':
          $ref: '#/components/responses/Unauthorized'
  /whatsapp/session/status:
    get:
      tags:
        - WhatsApp
      summary: Get session status
      description: Get current WhatsApp session status
      responses:
        '200':
          description: Session status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        enum: [connected, disconnected, waiting_for_qr, not_initialized, error]
                        example: "connected"
                      timestamp:
                        type: string
                        format: date-time
                        example: "2024-01-01T12:00:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
  /whatsapp/session/reconnect:
    post:
      tags:
        - WhatsApp
      summary: Reconnect session
      description: Manually reconnect WhatsApp session
      responses:
        '200':
          description: Session reconnected successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Session reconnected successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Validation failed"
              details:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Invalid or expired token"
...